ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-multipart==0.0.12
PyPDF2==3.0.1
python-docx==1.1.2
python-dateutil==2.9.0
